requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
PyPDF2
pdf2image
pytesseract
openai
